Added recipe show spec with recipe action helper
